requirements.txt
Django==1.8.2
django-braces==1.8.0
django-filter==0.10.0
django-oauth-toolkit==0.8.1
djangorestframework==3.1.2
Markdown==2.6.2
oauthlib==0.7.2
six==1.9.0
wheel==0.24.0
